data Model 리팩토링 & room 라이브러리 추가
