--- 2005/sangjinKIM93/newproject/app/src/main/java/com/sangjin/newproject/adapter/Movie.kt
@@ -1,16 +1,0 @@
-package com.sangjin.newproject.adapter
-
-data class ResponseData(
-    val items: List<Movie>
-)
-
-data class Movie(
-    val title: String,
-    val link: String,
-    val image: String,
-    val subtitle: String,
-    val pubDate: String,
-    val director: String,
-    val actor: String,
-    val userRating: String
-)
