+++ 2005/sangjinKIM93/newproject/app/src/main/java/com/sangjin/newproject/data/model/NaverMovieResponse.kt
@@ -1,0 +1,5 @@
+package com.sangjin.newproject.data.model
+
+data class NaverMovieResponse (
+    val items: List<Movie>
+)
