+++ 2005/sangjinKIM93/newproject/app/src/main/java/com/sangjin/newproject/data/repository/NaverMoviesRepository.kt
@@ -1,13 +1,13 @@
 package com.sangjin.newproject.data.repository
 
-import com.sangjin.newproject.adapter.Movie
+import com.sangjin.newproject.data.model.Movie
 
 interface NaverMoviesRepository {
 
     fun getNaverMovies(
         query: String,
         onSuccess: ((movies: List<Movie>) -> Unit),
-        onFailure: ((t:Throwable) -> Unit)
+        onFailure: ((t: Throwable) -> Unit)
     )
 
 }
