+++ 2005/sangjinKIM93/newproject/app/src/main/java/com/sangjin/newproject/data/source/remote/RemoteDataSource.kt
@@ -1,6 +1,6 @@
 package com.sangjin.newproject.data.source.remote
 
-import com.sangjin.newproject.adapter.Movie
+import com.sangjin.newproject.data.model.Movie
 
 interface RemoteDataSource {
